new file:   chksbc.sh 	new file:   date.sh 	new file:   deplst.sh         new file:   dialog.sh 	modified:   dspinstall.sh 	new file:   ex.sh 	new file:   examplbackup.sh 	new file:   examplesound.sh 	new file:   filechoose.sh 	new file:   imagelst.sh 	new file:   imgmv.sh 	new file:   makeitboot.sh 	new file:   menu.sh 	new file:   menu2.sh 	modified:   mvslb.sh 	new file:   rootmenu.sh 	new file:   sampled.sh 	new file:   sblist.sh 	new file:   test1.sh 	modified:   wholist.sh

+++ mvslb.sh
@@ -1,33 +1,40 @@
 #!/bin/bash
 NAME=$1
 DEFPATH="/home/dan/git/slackbuilds"
-FILEIN=$DEFPATH'/dspbuilds.txt'
+FILEIN=$DEFPATH'/slackbuilds.lst'
 TMPFILE='/tmp/dsp.tmp'
 CWD=$(pwd)
-WKGPATH='/home/dan/git/custom_slackbuilds/dspbuilds/'
-#echo $WKGPATH
-grep -A 9 'NAME: '$1 $FILEIN | cut -d " " -f2- > $TMPFILE
-exec < $TMPFILE
-read  NAME  
-read  LOCATION
-read  FILES
-read  VERSION
-read  DOWNLOAD
-read  DOWNLOAD_x86_64
-read  MD5SUM
-read  MD5SUM_x86_64
-read  REQUIRES
-read  SHORT
-rm $TMPFILE
-mkdir $WKGPATH/$LOCATION
-cd $WKGPATH/$LOCATION
-cp -vur $DEFPATH/$LOCATION/* ./
-sed -e s/SBo/dsp/g -e s/SlackBuild/dspbuild/g < $NAME.SlackBuild > $NAME.dspbuild
-rm $NAME.SlackBuild
-chmod +x $NAME.dspbuild
-cd $CWD
-./dspbuildlist.sh > dspbuilds.txt
-echo ______________________________________________
-echo $NAME
-echo Moved to $LOCATION
-echo ______________________________________________
+WKGPATH='/home/dan/git/custom_slackbuilds/dspbuilds'
+if [ -f $FILEIN ]; then
+    grep -A 9 'NAME: '$1 $FILEIN | cut -d " " -f2- > $TMPFILE
+    if [ $(grep -c 'NAME: '$1 $FILEIN) -lt 1 ]; then
+        echo ERROR
+        exit 69
+    else
+        exec < $TMPFILE
+        read  NAME  
+        read  LOCATION
+        read  FILES
+        read  VERSION
+        read  DOWNLOAD
+        read  DOWNLOAD_x86_64
+        read  MD5SUM
+        read  MD5SUM_x86_64
+        read  REQUIRES
+        read  SHORT
+        rm $TMPFILE
+        mkdir $WKGPATH/$LOCATION
+        cd $WKGPATH/$LOCATION
+        cp -vur $DEFPATH/$LOCATION/* ./
+        echo '#Script Modified by Dan Stein (email:dstein614@gmail.com) on '$(date) > $NAME.dspbuild
+        sed -e s/SBo/dsp/g -e s/SlackBuild/dspbuild/g < $NAME.SlackBuild >> $NAME.dspbuild
+        rm $NAME.SlackBuild
+        chmod +x $NAME.dspbuild
+        cd $CWD
+        ./dspbuildlist.sh > dspbuilds.txt
+        echo ______________________________________________
+        echo $NAME
+        echo Moved to $LOCATION
+        echo ______________________________________________
+    fi
+fi
